requirements.txt

python-telegram-bot==13.15
pytz==2025.2

